Add a commands to allow the `colorloop` effect
The API has a feature that allows the hub to send a sequence of colors (with transitions) to the bulbs that support it. This PR allows a user to invoke this effect on or off for a particular light.

```
User> hubot hue colorloop on light 1
Hubot> Setting light 1 colorloop to on

[..]

User> hubot hue colorloop off light 1
Hubot> Setting light 1 colorloop to off
```